feat: manually define templates from UI
- POST /api/templates accepts partial {name, config}; merges onto the
  default template (deep merge incl. per-level headings) so only the
  fields the user fills matter; font size supports floats (10.5pt)
- Frontend: 新建模板 button opens the format form in create mode;
  save routes to POST vs PUT by mode
- Tests: 33 total (manual partial config fills defaults)

+++ backend/app/models/schema.py
@@ -30,7 +30,7 @@ class PageConfig(BaseModel):
 class FontConfig(BaseModel):
     chinese: str = "宋体"
     english: str = "Times New Roman"
-    size: int = 12  # 磅
+    size: float = 12.0  # 磅（支持 10.5 等五号字）
     heading_font: str = "黑体"
     heading_english: str = "Arial"
 
@@ -47,7 +47,7 @@ class HeadingConfig(BaseModel):
     level: int
     font: str
     font_en: str
-    size: int
+    size: float
     bold: bool = True
     alignment: str = "left"
     space_before: float = 12  # pt
